+++ integration/securitypolicy_test.go
@@ -2,16 +2,20 @@ package integration
 
 import (
 	"context"
+	"os"
 	"testing"
 
 	"github.com/TykTechnologies/tyk-operator/api/v1alpha1"
+	"github.com/matryer/is"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	"github.com/TykTechnologies/tyk-operator/controllers"
 	tykClient "github.com/TykTechnologies/tyk-operator/pkg/client"
 	"github.com/TykTechnologies/tyk-operator/pkg/client/klient"
 	"github.com/TykTechnologies/tyk-operator/pkg/environmet"
-	"github.com/matryer/is"
+
 	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	"k8s.io/apimachinery/pkg/util/version"
 	ctrl "sigs.k8s.io/controller-runtime"
 	cr "sigs.k8s.io/controller-runtime/pkg/client"
@@ -22,6 +26,168 @@ import (
 	"sigs.k8s.io/e2e-framework/pkg/envconf"
 	"sigs.k8s.io/e2e-framework/pkg/features"
 )
+
+func TestSecurityPolicyStatusIsUpdated(t *testing.T) {
+	eval := is.New(t)
+
+	api1Name := "test-api-1-status"
+	api2Name := "test-api-2-status"
+	policyName := "test-policy"
+
+	mode := os.Getenv("TYK_MODE")
+	if mode == "ce" {
+		t.Skip("Skipping security policy test in CE mode")
+	}
+
+	policyCreate := features.New("SecurityPolicy status is updated").
+		Setup(func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
+			testNs, ok := ctx.Value(ctxNSKey).(string)
+			eval.True(ok)
+
+			_, err := createTestAPIDef(ctx, c, testNs, func(ad *v1alpha1.ApiDefinition) {
+				ad.Name = api1Name
+				ad.Spec.Name = api1Name
+				ad.Spec.Proxy.ListenPath = "/test-api-1"
+			})
+			eval.NoErr(err)
+
+			_, err = createTestAPIDef(ctx, c, testNs, func(ad *v1alpha1.ApiDefinition) {
+				ad.Name = api2Name
+				ad.Spec.Name = api2Name
+				ad.Spec.Proxy.ListenPath = "/test-api-2"
+			})
+			eval.NoErr(err)
+
+			// ensure API is created on Tyk before creating policy
+			apiDef := &v1alpha1.ApiDefinition{ObjectMeta: metav1.ObjectMeta{Name: api1Name, Namespace: testNs}}
+			err = waitForTykResourceCreation(c, apiDef)
+			eval.NoErr(err)
+
+			_, err = createTestPolicy(ctx, c, testNs, func(policy *v1alpha1.SecurityPolicy) {
+				policy.Name = policyName
+				policy.Spec.Name = policyName + testNs
+				policy.Spec.AccessRightsArray = []*v1alpha1.AccessDefinition{{Name: api1Name, Namespace: testNs}}
+			})
+			eval.NoErr(err)
+
+			pol := v1alpha1.SecurityPolicy{ObjectMeta: metav1.ObjectMeta{Name: policyName, Namespace: testNs}}
+			err = waitForTykResourceCreation(c, &pol)
+			eval.NoErr(err)
+
+			return ctx
+		}).Assess("validate links are created properly",
+		func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
+			var pol v1alpha1.SecurityPolicy
+			var api v1alpha1.ApiDefinition
+
+			testNs, ok := ctx.Value(ctxNSKey).(string)
+			eval.True(ok)
+
+			// check status of policy
+			err := c.Client().Resources().Get(ctx, policyName, testNs, &pol)
+			eval.NoErr(err)
+
+			eval.True(len(pol.Status.LinkedAPIs) != 0)
+			eval.Equal(pol.Status.LinkedAPIs[0].Name, api1Name)
+
+			// check status of ApiDefinition
+			err = c.Client().Resources().Get(ctx, api1Name, testNs, &api)
+			eval.NoErr(err)
+
+			eval.True(len(api.Status.LinkedByPolicies) != 0)
+			eval.Equal(api.Status.LinkedByPolicies[0].Name, policyName)
+
+			return ctx
+		}).Assess("Add new api in the access rights",
+		func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
+			var updatePolicy v1alpha1.SecurityPolicy
+			testNs, ok := ctx.Value(ctxNSKey).(string)
+			eval.True(ok)
+
+			err := c.Client().Resources().Get(ctx, policyName, testNs, &updatePolicy)
+			eval.NoErr(err)
+
+			updatePolicy.Spec.AccessRightsArray = append(updatePolicy.Spec.AccessRightsArray,
+				&v1alpha1.AccessDefinition{Name: api2Name, Namespace: testNs})
+
+			err = c.Client().Resources().Update(ctx, &updatePolicy)
+			eval.NoErr(err)
+
+			var pol v1alpha1.SecurityPolicy
+			pol.Name = policyName
+			pol.Namespace = testNs
+
+			// wait until status of policy is updated.
+			err = wait.For(conditions.New(c.Client().Resources()).ResourceMatch(&pol, func(object k8s.Object) bool {
+				pol, ok := object.(*v1alpha1.SecurityPolicy)
+				if !ok {
+					return false
+				}
+
+				if len(pol.Status.LinkedAPIs) == 2 {
+					return true
+				}
+
+				return false
+			}), wait.WithTimeout(defaultWaitTimeout), wait.WithInterval(defaultWaitInterval))
+			eval.NoErr(err)
+
+			eval.True(pol.Status.LinkedAPIs[0].Name == api2Name || pol.Status.LinkedAPIs[1].Name == api2Name)
+
+			var api v1alpha1.ApiDefinition
+			err = c.Client().Resources().Get(ctx, api2Name, testNs, &api)
+			eval.NoErr(err)
+
+			eval.True(len(api.Status.LinkedByPolicies) != 0)
+			eval.Equal(api.Status.LinkedByPolicies[0].Name, policyName)
+
+			return ctx
+		}).Assess("Delete access rights", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
+		var updatePolicy v1alpha1.SecurityPolicy
+
+		testNs, ok := ctx.Value(ctxNSKey).(string)
+		eval.True(ok)
+
+		err := c.Client().Resources().Get(ctx, policyName, testNs, &updatePolicy)
+		eval.NoErr(err)
+
+		updatePolicy.Spec.AccessRightsArray = nil
+
+		err = c.Client().Resources().Update(ctx, &updatePolicy)
+		eval.NoErr(err)
+
+		var pol v1alpha1.SecurityPolicy
+
+		pol.Name = policyName
+		pol.Namespace = testNs
+
+		err = wait.For(conditions.New(c.Client().Resources()).ResourceMatch(&pol, func(object k8s.Object) bool {
+			pol, ok := object.(*v1alpha1.SecurityPolicy)
+			if !ok {
+				return false
+			}
+
+			return pol.Status.LinkedAPIs == nil
+		}), wait.WithTimeout(defaultWaitTimeout), wait.WithInterval(defaultWaitInterval))
+		eval.NoErr(err)
+
+		var api v1alpha1.ApiDefinition
+
+		err = c.Client().Resources().Get(ctx, api1Name, testNs, &api)
+		eval.NoErr(err)
+
+		eval.True(api.Status.LinkedByPolicies == nil)
+
+		err = c.Client().Resources().Get(ctx, api2Name, testNs, &api)
+		eval.NoErr(err)
+
+		eval.True(api.Status.LinkedByPolicies == nil)
+
+		return ctx
+	}).Feature()
+
+	testenv.Test(t, policyCreate)
+}
 
 // minPolicyGwVersion represents the minimum Tyk Gateway version required to use Policy API.
 var minPolicyGwVersion = version.MustParseGeneric("v4.1.0")
@@ -366,7 +532,7 @@ func TestSecurityPolicy(t *testing.T) {
 				eval.True(ok)
 
 				var err error
-				apiDefCR, err = createTestAPIDef(ctx, testNs, nil, c)
+				apiDefCR, err = createTestAPIDef(ctx, c, testNs, nil)
 				eval.NoErr(err)
 
 				policyCR = v1alpha1.SecurityPolicy{
@@ -529,7 +695,7 @@ func TestSecurityPolicyForGraphQL(t *testing.T) {
 				testNs, ok := ctx.Value(ctxNSKey).(string)
 				eval.True(ok)
 
-				apiDefCR, err := createTestAPIDef(ctx, testNs, nil, c)
+				apiDefCR, err := createTestAPIDef(ctx, c, testNs, nil)
 				eval.NoErr(err)
 
 				err = waitForTykResourceCreation(c, apiDefCR)